fix(animatingnumbers): demo拆解与规范
