+++ src/packages/animatingnumbers/demos/h5/demo1.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { AnimatingNumbers, Cell } from '@nutui/nutui-react'
+
+const Demo1 = () => {
+  return <Cell title={<AnimatingNumbers.CountUp value="678.94" />} />
+}
+export default Demo1
